requirements.txt
youtube_dl==2020.5.3
spotipy==2.12.0
discord.py==1.3.3
mutagen==1.44.0
discord==1.0.1
python-dotenv==0.13.0
